Catch intptr_t to int conversion on 64-bit platforms.
Change the JNI generator to introduce a wrapper type
jni_int_wrapper which is used to prevent implicit
conversions which could be an intptr_t to int
conversion.

Change also includes fix (adding explicit case) for
a false positive which existings in the code base.

BUG=354405, 346626

Review URL: https://codereview.chromium.org/267893002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@268741 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: bb82c498a56d16388e92ce0d321e7641652a0ca0

+++ content/browser/android/content_video_view.cc
@@ -127,7 +127,8 @@ void ContentVideoView::UpdateMediaMetadata() {
   if (player && player->IsPlayerReady()) {
     Java_ContentVideoView_onUpdateMediaMetadata(
         env, content_video_view.obj(), player->GetVideoWidth(),
-        player->GetVideoHeight(), player->GetDuration().InMilliseconds(),
+        player->GetVideoHeight(),
+        static_cast<int>(player->GetDuration().InMilliseconds()),
         player->CanPause(),player->CanSeekForward(), player->CanSeekBackward());
   }
 }
